move default battle wrappers to battle_wrapper module
