+++ algobattle/battle.py
@@ -11,10 +11,6 @@ from pathlib import Path
 from typing import Literal
 import tomli
 from algobattle.battle_wrapper import BattleWrapper
-
-# for now we need to manually import the default wrappers to make sure they're initialized
-from algobattle.battle_wrappers.iterated import Iterated    # type: ignore # noqa: F401
-from algobattle.battle_wrappers.averaged import Averaged    # type: ignore # noqa: F401
 from algobattle.docker_util import DockerConfig, RunParameters
 
 from algobattle.match import MatchConfig, Match
