+++ tests/test_match.py
@@ -5,8 +5,7 @@ import logging
 from pathlib import Path
 
 from algobattle.battle import parse_cli_args, setup_logging
-from algobattle.battle_wrappers.iterated import Iterated
-from algobattle.battle_wrappers.averaged import Averaged
+from algobattle.battle_wrapper import Iterated, Averaged
 from algobattle.match import MatchConfig, Match
 from algobattle.team import Team, Matchup, TeamHandler, TeamInfo
 from algobattle.docker_util import DockerConfig, RunParameters, get_os_type
